+++ src/UnitTests/CalculatorSplitNumTests.cs
@@ -10,7 +10,7 @@ namespace TDDCalculator.UnitTests
         [InlineData("", 0)]
         [InlineData("1", 1)]
         [InlineData("10,2", 8)]
-        public void SplitNum_SplitNumTwoNumbers_WhenStringIsValid(string calculation, int expected)
+        public void SplitNum_SplitNumTwoNumbers_WhenStringIsValid(string calculation, float expected)
         {
             //Arrange
             var calc = new Calculator();
@@ -25,7 +25,7 @@ namespace TDDCalculator.UnitTests
         [Theory]
         [InlineData("60,15,10", 35)]
         [InlineData("140,45,35,20", 40)]
-        public void SplitNum_SplitNumAnyNumbers_WhenStringIsValid(string calculation, int expected)
+        public void SplitNum_SplitNumAnyNumbers_WhenStringIsValid(string calculation, float expected)
         {
             //Arrange
             var calc = new Calculator();
@@ -40,7 +40,7 @@ namespace TDDCalculator.UnitTests
         [Theory]
         [InlineData("8\n2,3", 3)]
         [InlineData("20\n2,3\n5, 5", 5)]
-        public void SplitNum_SplitNumUsingNewLineDelimiter_WhenStringIsValid(string calculation, int expected)
+        public void SplitNum_SplitNumUsingNewLineDelimiter_WhenStringIsValid(string calculation, float expected)
         {
             //Arrange
             var calc = new Calculator();
@@ -55,7 +55,7 @@ namespace TDDCalculator.UnitTests
         [Theory]
         [InlineData("8 SN 2 SN 3", 3)]
         [InlineData("140 SN 45 SN 35 SN 20", 40)]
-        public void SplitNum_SplitNumUsingSNDelimiter_WhenStringIsValid(string calculation, int expected)
+        public void SplitNum_SplitNumUsingSNDelimiter_WhenStringIsValid(string calculation, float expected)
         {
             //Arrange
             var calc = new Calculator();
