+++ src/UnitTests/CalculatorSubtractTests.cs
@@ -10,7 +10,7 @@ namespace TDDCalculator.UnitTests
         [InlineData("", 0)]
         [InlineData("1", 1)]
         [InlineData("3,2", 1)]
-        public void Subtract_SubtractTwoNumbers_WhenStringIsValid(string calculation, int expected)
+        public void Subtract_SubtractTwoNumbers_WhenStringIsValid(string calculation, float expected)
         {
             //Arrange
             var calc = new Calculator();
@@ -25,7 +25,7 @@ namespace TDDCalculator.UnitTests
         [Theory]
         [InlineData("6,1,1", 4)]
         [InlineData("10,2,3,5", 0)]
-        public void Subtract_SubtractAnyNumbers_WhenStringIsValid(string calculation, int expected)
+        public void Subtract_SubtractAnyNumbers_WhenStringIsValid(string calculation, float expected)
         {
             //Arrange
             var calc = new Calculator();
@@ -40,7 +40,7 @@ namespace TDDCalculator.UnitTests
         [Theory]
         [InlineData("8\n2,3", 3)]
         [InlineData("20\n2,3\n5, 5", 5)]
-        public void Subtract_SubtractUsingNewLineDelimiter_WhenStringIsValid(string calculation, int expected)
+        public void Subtract_SubtractUsingNewLineDelimiter_WhenStringIsValid(string calculation, float expected)
         {
             //Arrange
             var calc = new Calculator();
@@ -55,7 +55,7 @@ namespace TDDCalculator.UnitTests
         [Theory]
         [InlineData("8-2-3", 3)]
         [InlineData("20-2-3-5-5", 5)]
-        public void Subtract_SubtractUsingMinusDelimiter_WhenStringIsValid(string calculation, int expected)
+        public void Subtract_SubtractUsingMinusDelimiter_WhenStringIsValid(string calculation, float expected)
         {
             //Arrange
             var calc = new Calculator();
