+++ src/WebApi/Models/SplitEqRequest.cs
@@ -3,7 +3,7 @@ namespace TDDCalculator.WebApi.Models
 {
     public class SplitEqRequest
     {
-        public int Param1 { get; set; }
-        public int Param2 { get; set; }
+        public float Param1 { get; set; }
+        public float Param2 { get; set; }
     }
 }
